requirements.txt
# Core dependencies
requests>=2.31.0
dnspython>=2.4.0
jinja2>=3.1.0
colorama>=0.4.6
tqdm>=4.66.0

# Optional web enumeration
httpx>=0.25.0
aiohttp>=3.9.0

# Screenshot capture (optional)
playwright>=1.40.0
Pillow>=10.0.0

# Charting/visualization (optional)
matplotlib>=3.8.0
plotly>=5.18.0
kaleido>=0.2.1
pandas>=2.1.0

